requirements.txt
numpy==1.23.5
pandas==1.5.2
scikit-learn==1.2.1
streamlit==1.22.0
seaborn==0.12.2
matplotlib==3.6.2
Pillow==9.4.0
Flask==2.3.2
gunicorn==20.1.0
